@@ -1,65 +1,77 @@
-﻿// This Source Code Form is subject to the terms of the Mozilla Public
+// This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
+using System.Collections.Generic;
+using IdentityModel;
+using IdentityServer4.Models;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
 namespace IdentityServer4.Contrib.Membership.IdsvrDemo
 {
-    using Microsoft.AspNetCore.Builder;
-    using Microsoft.AspNetCore.Hosting;
-    using Microsoft.Extensions.Configuration;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Logging;
-    using System.Collections.Generic;
-    using IdentityModel;
-    using IdentityServer4.Models;
-    using Serilog;
-
     public class Startup
     {
-        public Startup(IHostingEnvironment env)
+        public Startup(IConfiguration configuration)
         {
-            var builder = new ConfigurationBuilder()
-                .SetBasePath(env.ContentRootPath)
-                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
-                .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true)
-                .AddEnvironmentVariables();
-            Configuration = builder.Build();
+            Configuration = configuration;
         }
 
-        public IConfigurationRoot Configuration { get; }
+        public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc();
+            services.AddControllersWithViews();
 
             services.AddIdentityServer()
-                    .AddDeveloperSigningCredential()
-                    .AddInMemoryClients(Clients.Get())
-                    .AddInMemoryIdentityResources(IdentityResources.Get())
-                    .AddMembershipService(new MembershipOptions
-                    {
-                        ConnectionString = "Data Source=localhost;Initial Catalog=Membership;Integrated Security=True",
-                        ApplicationName = "Test" 
-                    });
+                .AddDeveloperSigningCredential(persistKey: false)
+                .AddInMemoryClients(Clients.Get())
+                .AddInMemoryIdentityResources(IdentityResources.Get())
+                .AddMembershipService(new MembershipOptions
+                {
+                    ConnectionString = "Data Source=localhost;Initial Catalog=Membership;Integrated Security=True",
+                    ApplicationName = "Test",
+                    UseRoleProviderSource = true,
+                    MaxInvalidPasswordAttempts = 5
+                });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            loggerFactory.AddSerilog();
-
-            app.UseDeveloperExceptionPage();
-            app.UseStaticFiles();  
-                      
-            app.UseIdentityServer();
-            app.UseMvcWithDefaultRoute();
-        }
-
-        static class Clients
-        {
-            public static List<Client> Get()
+            if (env.IsDevelopment())
             {
-                return new List<Client>
+                app.UseDeveloperExceptionPage();
+            }
+            else
+            {
+                app.UseExceptionHandler("/Home/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                app.UseHsts();
+            }
+            app.UseStaticFiles();
+
+            app.UseRouting();
+            app.UseIdentityServer();
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
+                    name: "default",
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+            });
+        }
+    }
+
+    static class Clients
+    {
+        public static List<Client> Get()
+        {
+            return new List<Client>
                 {
                     new Client
                     {
@@ -80,7 +92,7 @@ namespace IdentityServer4.Contrib.Membership.IdsvrDemo
 
                         RedirectUris = new List<string>
                         {
-                            "http://localhost:5001/signin-oidc"
+                            "https://localhost:44321/signin-oidc"
                         },
 
                         AllowedScopes = new List<string>
@@ -92,33 +104,33 @@ namespace IdentityServer4.Contrib.Membership.IdsvrDemo
                             "ServiceStack.SelfHost",
                         },
 
-                        RequireConsent = false
+                        RequireConsent = false,
+                        RequirePkce = false
                     }
                 };
-            }
         }
+    }
 
-        static class IdentityResources
+    static class IdentityResources
+    {
+        public static List<IdentityResource> Get()
         {
-            public static List<IdentityResource> Get()
+            return new List<IdentityResource>
             {
-                return new List<IdentityResource>
+                new IdentityServer4.Models.IdentityResources.OpenId(),
+                new IdentityServer4.Models.IdentityResources.Profile(),
+                new IdentityServer4.Models.IdentityResources.Email(),
+                new IdentityResource
                 {
-                    new IdentityServer4.Models.IdentityResources.OpenId(),
-                    new IdentityServer4.Models.IdentityResources.Profile(),
-                    new IdentityServer4.Models.IdentityResources.Email(),
-                    new IdentityResource
+                    Name = "ServiceStack.SelfHost",
+                    Enabled = true,
+                    UserClaims = new List<string>
                     {
-                        Name = "ServiceStack.SelfHost",
-                        Enabled = true,
-                        UserClaims = new List<string>
-                        {
-                            JwtClaimTypes.Subject,
-                            JwtClaimTypes.Name
-                        }
+                        JwtClaimTypes.Subject,
+                        JwtClaimTypes.Name
                     }
-                };
-            }
+                }
+            };
         }
     }
 }
